refactor(batch): in main
